Må sette behandlingsresultatet når vedtaket besluttes - ikke når hele prosessen er ferdigstilt
