+++ src/main/kotlin/no/nav/familie/ef/sak/service/steg/FerdigstillBehandlingSteg.kt
@@ -1,9 +1,6 @@
 package no.nav.familie.ef.sak.service.steg
 
-import no.nav.familie.ef.sak.api.beregning.ResultatType
-import no.nav.familie.ef.sak.api.beregning.VedtakService
 import no.nav.familie.ef.sak.repository.domain.Behandling
-import no.nav.familie.ef.sak.repository.domain.BehandlingResultat
 import no.nav.familie.ef.sak.repository.domain.BehandlingStatus
 import no.nav.familie.ef.sak.repository.domain.BehandlingType
 import no.nav.familie.ef.sak.service.BehandlingService
@@ -13,13 +10,11 @@ import no.nav.familie.prosessering.domene.TaskRepository
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
-import java.util.UUID
 
 
 @Service
 class FerdigstillBehandlingSteg(private val behandlingService: BehandlingService,
-                                private val taskRepository: TaskRepository,
-                                private val vedtakService: VedtakService) : BehandlingSteg<Void?> {
+                                private val taskRepository: TaskRepository) : BehandlingSteg<Void?> {
 
     private val logger: Logger = LoggerFactory.getLogger(this::class.java)
 
@@ -28,7 +23,6 @@ class FerdigstillBehandlingSteg(private val behandlingService: BehandlingService
         behandlingService.oppdaterStatusPåBehandling(behandling.id, BehandlingStatus.FERDIGSTILT)
 
         if (behandling.type == BehandlingType.FØRSTEGANGSBEHANDLING || behandling.type == BehandlingType.REVURDERING) {
-            oppdaterResultatPåBehandling(behandling.id)
             taskRepository.save(PubliserVedtakshendelseTask.opprettTask(behandling.id))
             taskRepository.save(BehandlingsstatistikkTask.opprettFerdigTask(behandlingId = behandling.id))
         } else if (behandling.type == BehandlingType.BLANKETT || behandling.type == BehandlingType.TEKNISK_OPPHØR) {
@@ -41,12 +35,5 @@ class FerdigstillBehandlingSteg(private val behandlingService: BehandlingService
 
     override fun stegType(): StegType {
         return StegType.FERDIGSTILLE_BEHANDLING
-    }
-
-    fun oppdaterResultatPåBehandling(behandlingId: UUID) {
-        val vedtak = vedtakService.hentVedtak(behandlingId)
-        when (vedtak.resultatType) {
-            ResultatType.INNVILGE -> behandlingService.oppdaterResultatPåBehandling(behandlingId, BehandlingResultat.INNVILGET)
-        }
     }
 }
