+++ src/test/kotlin/no/nav/familie/ef/sak/service/steg/FerdigstillBehandlingStegTest.kt
@@ -3,42 +3,28 @@ package no.nav.familie.ef.sak.service.steg
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
-import no.nav.familie.ef.sak.api.beregning.ResultatType
-import no.nav.familie.ef.sak.api.beregning.VedtakService
 import no.nav.familie.ef.sak.no.nav.familie.ef.sak.repository.behandling
 import no.nav.familie.ef.sak.no.nav.familie.ef.sak.repository.fagsak
 import no.nav.familie.ef.sak.repository.domain.BehandlingType
-import no.nav.familie.ef.sak.repository.domain.Vedtak
 import no.nav.familie.ef.sak.service.BehandlingService
 import no.nav.familie.prosessering.domene.TaskRepository
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.catchThrowable
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import java.util.UUID
 
 internal class FerdigstillBehandlingStegTest {
 
     private val behandlingService = mockk<BehandlingService>(relaxed = true)
     private val taskRepository = mockk<TaskRepository>()
-    private val vedtakService = mockk<VedtakService>()
 
-    private val task = FerdigstillBehandlingSteg(behandlingService, taskRepository, vedtakService)
+    private val task = FerdigstillBehandlingSteg(behandlingService, taskRepository)
 
     private val fagsak = fagsak()
 
     @BeforeEach
     internal fun setUp() {
         every { taskRepository.save(any()) } answers { firstArg() }
-        every { vedtakService.hentVedtak(any()) } returns Vedtak(behandlingId = UUID.randomUUID(),
-                                                                 resultatType = ResultatType.INNVILGE,
-                                                                 periodeBegrunnelse = null,
-                                                                 inntektBegrunnelse = null,
-                                                                 avslåBegrunnelse = null,
-                                                                 perioder = null,
-                                                                 inntekter = null,
-                                                                 saksbehandlerIdent = null,
-                                                                 beslutterIdent = null)
     }
 
     @Test
